ml/requirements.txt
scikit-learn>=1.5
joblib>=1.4
numpy>=2.1
pandas>=2.2
psycopg2-binary>=2.9
scipy>=1.13
